Feat: Criando botão de curtir funcionário

+++ src/components/EmployeeCard/index.js
@@ -1,9 +1,9 @@
 import { useState } from 'react';
 import ImageWithFallBack from '../ImageWithFallback';
 import EmployeeCardStyle from './employee-card.css';
-import { AiFillCloseCircle } from "react-icons/ai";
+import { AiFillCloseCircle, AiOutlineHeart, AiFillHeart } from "react-icons/ai";
 
-const EmployeeCard = ({ id, name, role, srcImage, cardColor, onRemove }) => {
+const EmployeeCard = ({ id, name, role, srcImage, favorite, cardColor, onRemove, onFavorite }) => {
     
     const [isHovered, setIsHovered] = useState(false);
     const [isAnimating, setIsAnimating] = useState(false);
@@ -16,7 +16,10 @@ const EmployeeCard = ({ id, name, role, srcImage, cardColor, onRemove }) => {
         setIsHovered(false);
     };
 
-    const handleAnimationEnd = () => {
+    const handleAnimationEnd = (event) => {
+        if (event.animationName !== "DropCard")
+            return
+
         setIsAnimating(false);
         if (onRemove) {
             onRemove(id);
@@ -24,10 +27,14 @@ const EmployeeCard = ({ id, name, role, srcImage, cardColor, onRemove }) => {
     }
 
     const handleRemoveClick = (event) => {
-        event.stopPropagation();
         if (!isAnimating) {
             setIsAnimating(true);
         }
+    }
+
+    const handleFavoriteClick = () => {
+        onFavorite(id);
+        //onFavorite(id);
     }
 
     //Adiciona dinamicamente a calsse animate-click assim que o setIsAnimating(true) -> o componente irá realizar um update
@@ -55,6 +62,10 @@ const EmployeeCard = ({ id, name, role, srcImage, cardColor, onRemove }) => {
             <div className="e-card__footer">
                 <h4>{name}</h4>
                 <h5>{role}</h5>
+                <button className="favorite-button" onClick={handleFavoriteClick}>
+                    <AiFillHeart className={`favorite-button__icon ${favorite? "on" : "off"}`} color="#d51212" size={24} /> 
+                    <AiOutlineHeart className="favorite-button__icon" color="#3D3D3D" size={24}/>
+                </button>
             </div>
         </div>
     );
